pyproject.toml
[tool.poetry]
name = "indictranslate"
version = "0.1.0"
description = "Translate Indian languages to English and vice-versa"
authors = ["mukundesh <mukundesh@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
transformers = "^4.37.2"
indictranstokenizer = {git = "https://github.com/orgpedia/IndicTransTokenizer.git"}
more-itertools = "^10.2.0"
pysbd = "^0.3.4"
pyyaml = "^6.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
python = "^3.8"
transformers = "^4.37.2"
indictranstokenizer = {git = "https://github.com/orgpedia/IndicTransTokenizer.git"}
more-itertools = "^10.2.0"
pysbd = "^0.3.4"
pytest = "^8.0.0"
ruff = "^0.2.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
